fix(back) : refactor local game
