+++ front/game.css
@@ -14,7 +14,9 @@
     transform: translate(-50%, -50%);
 }
 
-
+.button-menu-quit {
+    
+}
 
 #game-menu {
     position: absolute;
