+++ front/index.html
@@ -302,7 +302,8 @@
                 <button id="game-menu-next" class="button-menu" type="button">Next Match</button>
                 <button id="game-menu-m3-quit" class="button-menu" type="button">Quit</button>
             </div>
-            
+
+
 
             <!-- in game banner -->
             <div id="game-menu-banner-phase" class="menu-element game-announce-base"></div>
